ajuster calcul d'absence,heure supp,max jours par mois

+++ ABRPOINT.Server/CalculService/HeureSupp/IHeureSuppService.cs
@@ -5,7 +5,7 @@ namespace ABRPOINT.Server.CalculService.HeureSupp
 {
     public interface IHeureSuppService
     {
-        public Task<int> CalculateHeureSupp(PresenceDto presence,Poste poste);
-        public Task<int> CalculateHeureSuppOptimise(PresenceDto presence,Poste poste);
+        public Task<double> CalculateHeureSupp(PresenceDto presence,Poste poste);
+        public Task<double> CalculateHeureSuppOptimise(PresenceDto presence,Poste poste);
     }
 }
